+++ src/main/java/cart/application/PaymentService.java
@@ -6,8 +6,8 @@ import cart.domain.point.MemberPoints;
 import cart.domain.point.Point;
 import cart.domain.point.SavePointPolicy;
 import cart.domain.point.UsedPoint;
-import cart.dto.PurchaseOrderItemRequest;
-import cart.dto.PurchaseOrderRequest;
+import cart.dto.purchaseorder.PurchaseOrderItemRequest;
+import cart.dto.purchaseorder.PurchaseOrderRequest;
 import cart.exception.PurchaseOrderException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
