+++ src/main/java/cart/dto/cartItem/CartItemRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.cartItem;
 
 public class CartItemRequest {
     private Long productId;
