+++ src/main/java/cart/dto/purchaseorder/request/PurchaseOrderItemRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.purchaseorder;
 
 public class PurchaseOrderItemRequest {
