+++ src/main/java/cart/ui/CartItemApiController.java
@@ -1,11 +1,10 @@
 package cart.ui;
 
 import cart.application.CartItemService;
-import cart.domain.CartItem;
 import cart.domain.Member;
-import cart.dto.CartItemQuantityUpdateRequest;
-import cart.dto.CartItemRequest;
-import cart.dto.CartItemResponse;
+import cart.dto.cartItem.CartItemQuantityUpdateRequest;
+import cart.dto.cartItem.CartItemRequest;
+import cart.dto.cartItem.CartItemResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
