+++ src/main/java/cart/ui/PurchaseOrderApiController.java
@@ -3,8 +3,8 @@ package cart.ui;
 import cart.application.PaymentService;
 import cart.application.PurchaseOrderService;
 import cart.domain.Member;
-import cart.dto.PurchaseOrderPageResponse;
-import cart.dto.PurchaseOrderRequest;
+import cart.dto.purchaseorder.PurchaseOrderRequest;
+import cart.dto.purchaseorder.response.PurchaseOrderPageResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
