+++ src/test/java/cart/application/PurchaseOrderServiceTest.java
@@ -3,8 +3,8 @@ package cart.application;
 import cart.dao.PurchaseOrderDao;
 import cart.dao.PurchaseOrderItemDao;
 import cart.domain.Member;
-import cart.dto.PurchaseOrderItemInfoResponse;
-import cart.dto.PurchaseOrderPageResponse;
+import cart.dto.purchaseorder.PurchaseOrderItemInfoResponse;
+import cart.dto.purchaseorder.response.PurchaseOrderPageResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
